+++ action/protocol/staking/stakebucketsbycandidate_test.go
@@ -1,4 +1,4 @@
-package action
+package staking
 
 import (
 	"encoding/hex"
@@ -17,7 +17,7 @@ func TestCallDataToStakeStateContextBucketsByCandidate(t *testing.T) {
 	req, err := CallDataToStakeStateContext(data)
 
 	r.Nil(err)
-	r.EqualValues("*action.BucketsByCandidateStateContext", reflect.TypeOf(req).String())
+	r.EqualValues("*staking.BucketsByCandidateStateContext", reflect.TypeOf(req).String())
 
 	method := &iotexapi.ReadStakingDataMethod{
 		Method: iotexapi.ReadStakingDataMethod_BUCKETS_BY_CANDIDATE,
